// src/models/VerificationStage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Reacon API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime.js';
/**
 * 
 * @export
 * @interface VerificationStage
 */
export interface VerificationStage {
    [key: string]: any | any;
    /**
     * 
     */
    cached?: boolean;
    /**
     * 
     */
    ips?: Array<string>;
    /**
     * 
     */
    mxHost?: string;
    /**
     * Known values: DNS, MX, SMTP. Clients preserve future values.
     */
    stage: string;
    /**
     * 
     */
    updatedAt: string;
}

/**
 * Check if a given object implements the VerificationStage interface.
 */
export function instanceOfVerificationStage(value: object): value is VerificationStage {
    if (!('stage' in value) || value['stage'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    return true;
}

export function VerificationStageFromJSON(json: any): VerificationStage {
    return VerificationStageFromJSONTyped(json, false);
}

export function VerificationStageFromJSONTyped(json: any, ignoreDiscriminator: boolean): VerificationStage {
    if (json == null) {
        return json;
    }
    return {
        
            ...json,
        'cached': json['cached'] == null ? undefined : json['cached'],
        'ips': json['ips'] == null ? undefined : json['ips'],
        'mxHost': json['mxHost'] == null ? undefined : json['mxHost'],
        'stage': json['stage'],
        'updatedAt': json['updatedAt'],
    };
}

export function VerificationStageToJSON(json: any): VerificationStage {
    return VerificationStageToJSONTyped(json, false);
}

export function VerificationStageToJSONTyped(value?: VerificationStage | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
            ...value,
        'cached': value['cached'],
        'ips': value['ips'],
        'mxHost': value['mxHost'],
        'stage': value['stage'],
        'updatedAt': value['updatedAt'],
    };
}