// src/models/VerificationProgress.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Reacon API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime.js';
/**
 * 
 * @export
 * @interface VerificationProgress
 */
export interface VerificationProgress {
    [key: string]: any | any;
    /**
     * 
     */
    capabilityEpoch?: number;
    /**
     * 
     */
    greylistAttempt?: number;
    /**
     * 
     */
    invalidationReason?: string;
    /**
     * 
     */
    mxFingerprint?: string;
    /**
     * 
     */
    nextRetryAt?: string;
    /**
     * 
     */
    preferredIp?: string;
    /**
     * 
     */
    reputationGeneration?: number;
    /**
     * 
     */
    requestId: string;
    /**
     * Known values: clean, dirty. Clients preserve future values.
     */
    requiredTier?: string;
    /**
     * Known values: queued, processing. Clients preserve future values.
     */
    state: string;
    /**
     * 
     */
    updatedAt: string;
}

/**
 * Check if a given object implements the VerificationProgress interface.
 */
export function instanceOfVerificationProgress(value: object): value is VerificationProgress {
    if (!('requestId' in value) || value['requestId'] === undefined) return false;
    if (!('state' in value) || value['state'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    return true;
}

export function VerificationProgressFromJSON(json: any): VerificationProgress {
    return VerificationProgressFromJSONTyped(json, false);
}

export function VerificationProgressFromJSONTyped(json: any, ignoreDiscriminator: boolean): VerificationProgress {
    if (json == null) {
        return json;
    }
    return {
        
            ...json,
        'capabilityEpoch': json['capabilityEpoch'] == null ? undefined : json['capabilityEpoch'],
        'greylistAttempt': json['greylistAttempt'] == null ? undefined : json['greylistAttempt'],
        'invalidationReason': json['invalidationReason'] == null ? undefined : json['invalidationReason'],
        'mxFingerprint': json['mxFingerprint'] == null ? undefined : json['mxFingerprint'],
        'nextRetryAt': json['nextRetryAt'] == null ? undefined : json['nextRetryAt'],
        'preferredIp': json['preferredIp'] == null ? undefined : json['preferredIp'],
        'reputationGeneration': json['reputationGeneration'] == null ? undefined : json['reputationGeneration'],
        'requestId': json['requestId'],
        'requiredTier': json['requiredTier'] == null ? undefined : json['requiredTier'],
        'state': json['state'],
        'updatedAt': json['updatedAt'],
    };
}

export function VerificationProgressToJSON(json: any): VerificationProgress {
    return VerificationProgressToJSONTyped(json, false);
}

export function VerificationProgressToJSONTyped(value?: VerificationProgress | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
            ...value,
        'capabilityEpoch': value['capabilityEpoch'],
        'greylistAttempt': value['greylistAttempt'],
        'invalidationReason': value['invalidationReason'],
        'mxFingerprint': value['mxFingerprint'],
        'nextRetryAt': value['nextRetryAt'],
        'preferredIp': value['preferredIp'],
        'reputationGeneration': value['reputationGeneration'],
        'requestId': value['requestId'],
        'requiredTier': value['requiredTier'],
        'state': value['state'],
        'updatedAt': value['updatedAt'],
    };
}