// src/models/VerificationFinal.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Reacon API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime.js';
import type { VerificationResult } from './VerificationResult.js';
import {
    VerificationResultFromJSON,
    VerificationResultFromJSONTyped,
    VerificationResultToJSON,
    VerificationResultToJSONTyped,
} from './VerificationResult.js';

/**
 * 
 * @export
 * @interface VerificationFinal
 */
export interface VerificationFinal {
    [key: string]: any | any;
    /**
     * 
     */
    result: VerificationResult;
    /**
     * 
     */
    updatedAt: string;
}

/**
 * Check if a given object implements the VerificationFinal interface.
 */
export function instanceOfVerificationFinal(value: object): value is VerificationFinal {
    if (!('result' in value) || value['result'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    return true;
}

export function VerificationFinalFromJSON(json: any): VerificationFinal {
    return VerificationFinalFromJSONTyped(json, false);
}

export function VerificationFinalFromJSONTyped(json: any, ignoreDiscriminator: boolean): VerificationFinal {
    if (json == null) {
        return json;
    }
    return {
        
            ...json,
        'result': VerificationResultFromJSON(json['result']),
        'updatedAt': json['updatedAt'],
    };
}

export function VerificationFinalToJSON(json: any): VerificationFinal {
    return VerificationFinalToJSONTyped(json, false);
}

export function VerificationFinalToJSONTyped(value?: VerificationFinal | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
            ...value,
        'result': VerificationResultToJSON(value['result']),
        'updatedAt': value['updatedAt'],
    };
}