// src/models/VerificationStreamError.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Reacon API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime.js';
/**
 * 
 * @export
 * @interface VerificationStreamError
 */
export interface VerificationStreamError {
    [key: string]: any | any;
    /**
     * 
     */
    code: string;
    /**
     * 
     */
    error: string;
    /**
     * 
     */
    remainingCredits?: number;
    /**
     * 
     */
    status?: number;
    /**
     * 
     */
    updatedAt: string;
}

/**
 * Check if a given object implements the VerificationStreamError interface.
 */
export function instanceOfVerificationStreamError(value: object): value is VerificationStreamError {
    if (!('code' in value) || value['code'] === undefined) return false;
    if (!('error' in value) || value['error'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    return true;
}

export function VerificationStreamErrorFromJSON(json: any): VerificationStreamError {
    return VerificationStreamErrorFromJSONTyped(json, false);
}

export function VerificationStreamErrorFromJSONTyped(json: any, ignoreDiscriminator: boolean): VerificationStreamError {
    if (json == null) {
        return json;
    }
    return {
        
            ...json,
        'code': json['code'],
        'error': json['error'],
        'remainingCredits': json['remainingCredits'] == null ? undefined : json['remainingCredits'],
        'status': json['status'] == null ? undefined : json['status'],
        'updatedAt': json['updatedAt'],
    };
}

export function VerificationStreamErrorToJSON(json: any): VerificationStreamError {
    return VerificationStreamErrorToJSONTyped(json, false);
}

export function VerificationStreamErrorToJSONTyped(value?: VerificationStreamError | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
            ...value,
        'code': value['code'],
        'error': value['error'],
        'remainingCredits': value['remainingCredits'],
        'status': value['status'],
        'updatedAt': value['updatedAt'],
    };
}